pyproject.toml
[project]
name = "qaguruadvancedpython"
version = "0.1.0"
description = ""
authors = [
    {name = "Максим Андреев",email = "mp.andreev@interfax.ru"}
]
readme = "README.md"
requires-python = ">=3.13, <4.0"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "starlette (>=0.46.2,<0.47.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "requests (>=2.32.3,<3.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "fastapi-pagination (>=0.13.1,<0.14.0)",
    "pydantic[email] (>=2.11.5,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
